Add test with examples of well-formed tax IDs
This is the next logical test case, because it forces us to change the
implementation. Now that we have an accepting test case, we can add
more rejecting test cases which will force us to add/change logic
step by step as we implement more rules. Otherwise we could add more
rejecting tests without having them fail first, since rejecting test
cases assert the same behavior (reject the tax ID input, i.e., throw
an exception) and differ only in the test data. In general we should
try to add the next test case that fails without having to change the
implementation, and that forces us to change the implementation in
order to make the test pass.

By using a parameterized test, we can test multiple values at once
without having to add redundant test methods. This keeps the test class
simple and the rules easy to understand. It also reduces the likelihood
of false positives compared to testing only a single value.
In this case, some examples of well-formed tax IDs are provided with
the requirements (i.e., the kata instructions). If that weren't the
case, we would have to create our own by manually creating example
strings that fulfill the given rules of well-formed tax IDs, or ask
domain experts.

Note that for now we simply return null when the input string is
well-formed. This is fine, since we have not yet described a case
of actually using the return value. This will be another test case
in the future, and the return value is simply not specified and
therefore not ready to use yet.