A well-formed tax ID must be exactly 11 chars long
We check the length next, although checking one of the other rules
first would also work. The length check is just a very simple test
case.

Note that we are intentionally choosing the term "well-formed" (and
its counterpart, "ill-formed") because the term "valid" could mean
"assigned to a tax payer". It's important to be precise with our
naming, even if the names might seem very similar at first glance.