Refactor and simplify tests (step 1)
(Normally we would not split these 2 steps into separate commits, but
this is to illustrate the process.)

We create another test (a parameterized one) that represents exactly
the 3 test cases above it. It is important to regularly check if we
can simplify our tests to make them easier to understand and maintain.
Of course we don't always want to just combine multiple tests into one,
but sometimes we find that combining two test cases results in one that
is as easy, if not easier, to understand. It's important that the test
name still accurately describes the whole test case and makes it
obvious why each of the test values is used. And with a parameterized
test, we can very easily add additional test cases without having to
add more test methods.

Note that the name "...exactly_11_digits_and_nothing_else" is very
explicit. Just saying "...exactly_11_digits" could mean "11 digits,
but also whitespace and/or other characters".

In the next commit (and normally in the same commit) we can remove the
now obsolete 3 test cases for empty string, string length, and digits.