Simplify tax ID validation using regex
Often regular expressions can be quite hard to read, but in some cases
a very simple regex can be easy to understand and at the same time make
the implementation of validation rules much simpler, like in this case.
And if the tests are already there, we can safely refactor the
implementation.